Move to logging.py
